% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWeight.R
\name{getWeight}
\alias{getWeight}
\title{Compute the weights for neighborhood.}
\usage{
getWeight(matrix, bandwidth, weightmethod = "gaussian")
}
\arguments{
\item{matrix}{data frame (origin, destination, travel_time structure) to be considered for weighting}

\item{bandwidth}{numeric bandwidth selected to perform neighborhood (same unit of distance)}

\item{weightmethod}{numeric to choose the method to be used: gaussian , bi-square and moving window. Defaults to gaussian}
}
\value{
data frame with weight value for internal use (population intensity)
}
\description{
Compute the weights for neighborhood.
}
