% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalIndexH.R
\name{globalIndexH}
\alias{globalIndexH}
\title{Function to compute global index H}
\usage{
globalIndexH(self)
}
\arguments{
\item{self}{list with population intensity calculated (result from popIntensity)}
}
\value{
: values with global index for each group.
}
\description{
Function to compute global index H
}
