% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalMoran.R
\name{globalMoran}
\alias{globalMoran}
\title{Processes the global Moran's I bivariate}
\usage{
globalMoran(shp, x = NULL, y = NULL, W = NULL)
}
\arguments{
\item{shp}{shapefile or simple feature to process the neighbourhood. Expected structure: id, x, y(optional)}

\item{x}{numeric vector with values of variable 1. Defaults to NULL.}

\item{y}{numeric vector with values of variable 2. Defaults to NULL (not bivariate).}

\item{W}{numeric matrix of neighbourhood matrix from \code{\link{nbMatrix}}. Defaults to NULL.}
}
\value{
the global value of Moran's I
}
\description{
Processes the global Moran's I bivariate
}
