% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moran.R
\name{globalMoranSim}
\alias{globalMoranSim}
\title{Evaluates the p-value of the Moran's I}
\usage{
globalMoranSim(shp, x = NULL, y = NULL, W = NULL, nsims = 999)
}
\arguments{
\item{shp}{shapefile or simple feature to process the neighbourhood. Expected structure: id, x, y(optional)}

\item{x}{numeric vector with values of variable 1. Defaults to NULL.}

\item{y}{numeric vector with values of variable 2. Defaults to NULL (not bivariate).}

\item{W}{numeric matrix of neighbourhood matrix from \code{\link{nbMatrix}}. Defaults to NULL.}

\item{nsims}{numeric value of number of simulations. Defaults to 999.}
}
\value{
global_sims (global Moran's I simulations)
}
\description{
Evaluates the p-value of the Moran's I
}
