% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segreg_global.R
\name{globalSegreg}
\alias{globalSegreg}
\title{Summary of global segregation indexes}
\usage{
globalSegreg(self, digits = 5)
}
\arguments{
\item{self}{list with population intensity calculated (result from \code{popIntensity_matrix})}

\item{digits}{numeric number of digits of display. Defaults to 5}
}
\value{
list with calculated global indexes with the following items
  diss: numeric dissimilarity index
  ent: numeric entropy
  iH: numeric index H
  exp: data frame exposure indexes for group m with group n (size n x n)
}
\description{
Summary of global segregation indexes
}
