% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segreg_prep.R
\name{idProcess}
\alias{idProcess}
\title{Processes the ids of data frame 1 and data frame 2}
\usage{
idProcess(df1, df2)
}
\arguments{
\item{df1}{A data frame with ids to be processed.}

\item{df2}{A data frame with ids to be processed.}
}
\value{
A list with the two data frames processed with common ids and with the same order.
#' @author Beatriz Moura dos Santos
}
\description{
Processes the ids of data frame 1 and data frame 2
}
