% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segreg_local.R
\name{localDissimilarity}
\alias{localDissimilarity}
\title{Compute local dissimilarity for all groups.}
\usage{
localDissimilarity(self)
}
\arguments{
\item{self}{list with population intensity calculated (result from popIntensity)}
}
\value{
array with dissimilarity for all groups, size of number of localities
}
\description{
Compute local dissimilarity for all groups.
}
