% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segreg_local.R
\name{localExposure}
\alias{localExposure}
\title{This function computes the local exposure index of group m to group n.
in situations where m=n, then the result is the isolation index.}
\usage{
localExposure(self)
}
\arguments{
\item{self}{list with population intensity calculated (result from popIntensity)}
}
\value{
data frame with local exposure indexes, size n*n (n=number of groups) x number of localities
}
\description{
This function computes the local exposure index of group m to group n.
in situations where m=n, then the result is the isolation index.
}
