% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localIndexH.R
\name{localIndexH}
\alias{localIndexH}
\title{This function computes the local entropy index H for all localities.}
\usage{
localIndexH(self)
}
\arguments{
\item{self}{list with population intensity calculated (result from popIntensity)}
}
\value{
array with scores for n groups (number of groups)
}
\description{
This function computes the local entropy index H for all localities.
}
