% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbMatrix.R
\name{nbMatrix}
\alias{nbMatrix}
\title{Processes the neighbourhood matrix}
\usage{
nbMatrix(shp)
}
\arguments{
\item{shp}{shapefile or simple feature of locations}
}
\value{
numeric binary neighbourhood matrix
}
\description{
Processes the neighbourhood matrix
}
