% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segreg_prep.R
\name{popIntensity}
\alias{popIntensity}
\title{Computes the local population intensity for all groups based on a time/distance matrix.}
\usage{
popIntensity(matrix, pop, bandwidth = 60, weightmethod = "gaussian")
}
\arguments{
\item{matrix}{A data frame with travel time/distance matrix (origin, destination, travel_time).}

\item{pop}{A data frame structured as id and columns with number of population for each group.}

\item{bandwidth}{A numeric bandwidth for neighborhood (same unit of matrix). Defaults to 60 (minutes).}

\item{weightmethod}{A string object to choose the method to be used: gaussian , bi-square and moving window. 
Defaults to gaussian.}
}
\value{
A data frame with population intensity for all groups.
}
\description{
Computes the local population intensity for all groups based on a time/distance matrix.
}
\references{
Smith, Goodchild & Longley (2018). 
   \emph{Geospatial Analysis. A Comprehensive Guide to PrinciplesTechniques and Software Tools}. 
   \url{https://www.spatialanalysisonline.com/HTML/index.html}
}
\author{
Beatriz Moura dos Santos
}
