% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential.R
\name{potential}
\alias{potential}
\title{Processes the potential (or gravity) job accessibility.}
\usage{
potential(betas, oportunities, matrix)
}
\arguments{
\item{betas}{numeric 'willing' decay factor}

\item{oportunities}{data frame with at least id and job fields}

\item{matrix}{data frame with origin, destination and travel_time structure}
}
\value{
data frame with origin and acc
}
\description{
Uses the exponential function to weight the oportunities.
}
