% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potGaussian.R
\name{potGaussian}
\alias{potGaussian}
\title{Processes the potential accessibility with Gaussian function}
\usage{
potGaussian(cutoff, oportunities, matrix)
}
\arguments{
\item{cutoff}{numeric threshold of cumulative accessibility}

\item{oportunities}{data frame with at least id and job fields}

\item{matrix}{data frame with origin, destination and travel_time structure}
}
\value{
data frame with origin, acc and threshold
}
\description{
Processes the potential accessibility with Gaussian function
}
