% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAttributesFile.R
\name{readAttributesFile}
\alias{readAttributesFile}
\title{This function reads the csv file and populate the class's attributes.
Data has to be exactly in the following format or results will be wrong:
area id,  x_coord, y_coord, population of G1, pop G2, pop G3, ..., pop Gn (n groups)}
\usage{
readAttributesFile(filepath)
}
\arguments{
\item{filepath}{string path with file to be read}
}
\value{
list with:
   attribute matrix: raw data,
   location: x_coord, y_coord,
   pop: population of each group,
   n_group: number of groups,
   n_location: number of tracts,
   pop_sum: total population of each tract,
   tract_id: tract id
}
\description{
This function reads the csv file and populate the class's attributes.
Data has to be exactly in the following format or results will be wrong:
area id,  x_coord, y_coord, population of G1, pop G2, pop G3, ..., pop Gn (n groups)
}
