% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moran.R
\name{simulaMoran}
\alias{simulaMoran}
\title{Evaluates the p-value of the Moran's I}
\usage{
simulaMoran(x, y = NULL, W = NULL, shp = NULL, nsims = 999)
}
\arguments{
\item{x}{numeric vector with values of variable 1. Defaults to NULL.}

\item{y}{numeric vector with values of variable 2. Defaults to NULL (not bivariate).}

\item{W}{numeric matrix of neighbourhood matrix from \code{\link{nbMatrix}}. Defaults to NULL.}

\item{shp}{shapefile or simple feature to process the neighbourhood. Expected structure: id, x, y(optional)}

\item{nsims}{numeric value of number of simulations. Defaults to 999.}
}
\value{
list with global_sim (global Moran's I simulations) and 
        local_sim (local Moran's I simulations)and lagged (neighbours' mean)
}
\description{
Evaluates the p-value of the Moran's I
}
