% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessibility_functions.R
\name{twosfca}
\alias{twosfca}
\title{Processes the jobs 2SFCA accessibility}
\usage{
twosfca(cutoff, oportunities, matrix, prestep = TRUE)
}
\arguments{
\item{cutoff}{numeric threshold of cumulative accessibility. It must be in the same unit of the matrix}

\item{oportunities}{data frame with at least id and job fields}

\item{matrix}{data frame with origin, destination and travel_time structure}

\item{prestep}{logical parameter. Defaults to TRUE to process the prestep 
that divides the population by the number of counts in the catchment area of the first step}
}
\value{
data frame with origin, acc and threshold
}
\description{
Processes the jobs 2SFCA accessibility
}
