% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessibility_functions.R
\name{cumop}
\alias{cumop}
\title{Processes the job cumulative accessibility}
\usage{
cumop(cutoff, oportunities, matrix)
}
\arguments{
\item{cutoff}{numeric threshold of cumulative accessibility. It must be in the same unit of the matrix}

\item{oportunities}{data frame with at least id and job fields}

\item{matrix}{data frame with origin, destination and travel_time structure}
}
\value{
data frame with origin, acc and threshold
}
\description{
Processes the job cumulative accessibility
}
\details{
The job cumulative accessibility adds up the number of oportunities inside the threshold.
}
