% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalDissimilarity.R
\name{globalDissimilarity}
\alias{globalDissimilarity}
\title{This function call local dissimilarity and compute the sum from individual values.}
\usage{
globalDissimilarity(self)
}
\arguments{
\item{self}{list with population intensity calculated (result from popIntensity)}
}
\value{
display global value of dissimilarity
}
\description{
This function call local dissimilarity and compute the sum from individual values.
}
