% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segreg_global.R
\name{globalEntropy}
\alias{globalEntropy}
\title{Computes the global entropy score E (diversity).}
\usage{
globalEntropy(self)
}
\arguments{
\item{self}{list with population intensity calculated (result from popIntensity)}
}
\value{
diversity score
}
\description{
Computes the global entropy score E (diversity).
}
