% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segreg_global.R
\name{globalEntropy}
\alias{globalEntropy}
\title{Computes the global entropy score E (diversity).}
\usage{
globalEntropy(pop, pop_intensity)
}
\arguments{
\item{pop}{A data frame with id, group, and number of population.}

\item{pop_intensity}{A data frame with id and population intensity for all groups. Result from \code{\link{popIntensity}}.}
}
\value{
Display the diversity score.
}
\description{
Computes the global entropy score E (diversity).
}
\references{
Iceland (2004. 
   The multigroup entropy index (also known as Theil’s H or the information theory index). 
   \emph{US Census Bureau}, Retrieved July, 31, 2006.

Sousa (2017). Segregation Metrics. 
   \url{https://github.com/sandrofsousa/Resolution/blob/master/Pysegreg/segregationMetrics.py}.
}
\author{
Beatriz Moura dos Santos
}
