% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segreg_global.R
\name{globalExposure}
\alias{globalExposure}
\title{Computes the global exposure measures}
\usage{
globalExposure(self)
}
\arguments{
\item{self}{list with population intensity calculated (result from popIntensity)}
}
\value{
data frame of global exposure result, size n x n (n=number of groups)
}
\description{
This function call local exposure function and sum the results for the global index.
}
