% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segreg_global.R
\name{globalIndexH}
\alias{globalIndexH}
\title{Computes global index H}
\usage{
globalIndexH(self)
}
\arguments{
\item{self}{list with population intensity calculated (result from popIntensity)}
}
\value{
: values with global index for each group.
}
\description{
Computes global index H
}
