% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segreg_global.R
\name{globalSegreg}
\alias{globalSegreg}
\title{Summary of global segregation indexes}
\usage{
globalSegreg(pop, pop_intensity, digits = 5)
}
\arguments{
\item{pop}{A data frame with id, group, and number of population.}

\item{pop_intensity}{data frame with id and population intensity for all groups. Result from \code{\link{popIntensity}}.}

\item{digits}{A numeric object of number of digits to display. Defaults to 5.}
}
\value{
list with calculated global indexes with the following items:
  \describe{
    \item{diss} {Numeric dissimilarity index.}
    \item{ent} {Numeric entropy.}
    \item{iH} {Numeric index H.}
    \item{exp} {Data frame exposure indexes for group m with group n (size n x n).}
  }
}
\description{
Summary of global segregation indexes
}
