% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LISAmaps.R
\name{LISAmaps}
\alias{LISAmaps}
\title{Processes the local simulations and local Moran to produce 
LISA map. IT STILL CAN'T PROCESS THE BIVARIATE}
\usage{
LISAmaps(shp, x = NULL, y = NULL, W = NULL, local_sims = NULL,
  local_moran = NULL)
}
\arguments{
\item{local_sims}{numeric vector with local simulations of Moran (result from \code{\link{localMoranSim}}}

\item{local_moran}{numeric vector with local values of Moran's I (result from \code{\link{localMoran}}}
}
\value{
sf with localPseudo (p-value of local), pvalue (p-value classified), patterns (cluster LISA map input)
}
\description{
Processes the local simulations and local Moran to produce 
LISA map. IT STILL CAN'T PROCESS THE BIVARIATE
}
