% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localEntropy.R
\name{localEntropy}
\alias{localEntropy}
\title{This function computes the local entropy score for a unit area Ei (diversity).}
\usage{
localEntropy(self)
}
\arguments{
\item{self}{list with population intensity calculated (result from popIntensity)}
}
\value{
array with local indices of entropy
}
\description{
This function computes the local entropy score for a unit area Ei (diversity).
}
