% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segreg_local.R
\name{localExposure}
\alias{localExposure}
\title{Computes the local exposure index of group m to group n.
in situations where m=n, then the result is the isolation index.}
\usage{
localExposure(pop, pop_intensity)
}
\arguments{
\item{pop}{A data frame with id, group, and number of population.}

\item{pop_intensity}{A data frame with id and population intensity for all groups. Result from \code{\link{popIntensity}}.}
}
\value{
data frame with local exposure indexes, size number of localities x n^2 (n=number of groups).
}
\description{
Computes the local exposure index of group m to group n.
in situations where m=n, then the result is the isolation index.
}
\references{
Feitosa, Camara, Monteiro, Koschitzi & Silva (2007). 
   Global and local spatial indices of urban segregation. 
   \emph{International Journal of Geographical Information Science}, 21(3), 299-323.

Sousa (2017). Segregation Metrics. 
   \url{https://github.com/sandrofsousa/Resolution/blob/master/Pysegreg/segregationMetrics.py}.
}
\author{
Beatriz Moura dos Santos
}
