% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segreg_prep.R
\name{localIntensity}
\alias{localIntensity}
\title{Processes the local intensity for each locality}
\usage{
localIntensity(matrix, pop)
}
\arguments{
\item{matrix}{data frame (origin, destination, weight)}

\item{pop}{data frame (id, group, population structure)}
}
\value{
data frame with local intensity for each group
}
\description{
Processes the local intensity for each locality
}
