% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localSegreg.R
\name{localSegreg}
\alias{localSegreg}
\title{Summary of local segregation indexes}
\usage{
localSegreg(self)
}
\arguments{
\item{self}{list with population intensity calculated (result from popIntensity)}
}
\value{
data frame with calculated indexes for each tract with the following columns
  id: tract id
  diss: dissimilarity index
  ent: entropy
  iH: index H
  exp_m_n: exposure inde for group m with group n
}
\description{
Summary of local segregation indexes
}
