% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segreg_prep.R
\name{popIntensity_matrix}
\alias{popIntensity_matrix}
\title{This function calculate the local population intensity for all groups based on a time matrix.}
\usage{
popIntensity_matrix(matrix, pop, bandwidth = 60,
  weightmethod = "gaussian")
}
\arguments{
\item{matrix}{data frame (origin, destination, travel_time)}

\item{pop}{data frame (id, group, population structure)}

\item{bandwidth}{numeric bandwidth for neighborhood (same unit of matrix). Defaults to 60 (minutes)}

\item{weightmethod}{numeric to choose the method to be used: gaussian , bi-square and moving window.
Defaults to gaussian}
}
\value{
list with:
   attribute matrix: raw data,
   location: x_coord, y_coord,
   pop: population of each group,
   n_group: number of groups,
   n_location: number of tracts,
   pop_sum: total population of each tract,
   tract_id: tract id,
   locality: population intensity for all groups.
}
\description{
This function calculate the local population intensity for all groups based on a time matrix.
}
