% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessibility_functions.R
\name{potential}
\alias{potential}
\title{Processes the potential (or gravity) job accessibility}
\usage{
potential(oportunities, matrix, type = "exponential", betas = NULL,
  cutoff = NULL)
}
\arguments{
\item{oportunities}{data frame with at least id and job fields}

\item{matrix}{data frame with origin, destination and travel_time structure}

\item{type}{string to choose between exponential and gaussian function to decay the oportunities. Defaults to exponential.}

\item{betas}{numeric 'willing' decay factor. It must be defined when type "exponential" is chosen. Defaults to NULL}

\item{cutoff}{numeric threshold. It must be in the same unit of the matrix and it must be defined when type "gaussian" is chosen. Defaults to NULL}
}
\value{
data frame with origin and acc
}
\description{
Uses the exponential function to weight the oportunities
}
